+++ apps/scan/frontend/src/components/TotalAmuletBalance.tsx
@@ -1,26 +1,29 @@
 import BigNumber from 'bignumber.js';
 import { ErrorDisplay, Loading } from 'common-frontend';
-import { useTotalAmuletBalance } from 'common-frontend/scan-api';
+import { useAmuletPrice, useTotalAmuletBalance } from 'common-frontend/scan-api';
 
 import AmountSummary from './AmountSummary';
 
 export const TotalAmuletBalance: React.FC = () => {
   const totalAmuletBalanceQuery = useTotalAmuletBalance();
+  const amuletPriceQuery = useAmuletPrice();
 
-  switch (totalAmuletBalanceQuery.status) {
-    case 'loading':
-      return <Loading />;
-    case 'error':
-      return <ErrorDisplay message={'Could not retrieve total amulet balance'} />;
-    case 'success':
-      return (
-        <AmountSummary
-          title="Total Canton Coin Balance"
-          amount={new BigNumber(totalAmuletBalanceQuery.data.total_balance)}
-          idCC="total-amulet-balance-cc"
-        />
-      );
-  }
+  const isLoading = totalAmuletBalanceQuery.isLoading || amuletPriceQuery.isLoading;
+  const isError = totalAmuletBalanceQuery.isError || amuletPriceQuery.isError;
+
+  return isLoading ? (
+    <Loading />
+  ) : isError ? (
+    <ErrorDisplay message={'Could not retrieve total amulet balance or amulet price'} />
+  ) : (
+    <AmountSummary
+      title="Total Canton Coin Balance"
+      amount={new BigNumber(totalAmuletBalanceQuery.data.total_balance)}
+      idCC="total-amulet-balance-cc"
+      idUSD="total-amulet-balance-usd"
+      amuletPrice={amuletPriceQuery.data}
+    />
+  );
 };
 
 export default TotalAmuletBalance;